refactor(runtime,instance): split large files and simplify state management
Reduce cognitive load in the instance controller (~7K LOC) and node
runtime (~3.1K LOC) through six incremental refactors:

1. Split resources.go by node type — external ref/collection processing
   moved to resources_external.go, collection processing moved to
   resources_collection.go. Zero behavioral change.

2. Split node.go by concern — context building to node_context.go,
   readiness checks to node_readiness.go, resolution logic to
   node_resolve.go. Zero behavioral change.

3. Consolidate expression caching — add EvalCached() on
   expressionEvaluationState, making evalBoolExpr/evalListExpr thin
   type-assertion wrappers and eliminating duplicated cache-check-and-set
   across 4 locations.

4. Make resolution mode explicit — add resolveMode enum and unify
   GetDesired/GetDesiredIdentity into a shared resolve(mode) method
   that encodes which vars to resolve, whether to check dep readiness,
   and whether to cache results.

5. Return NodeState as values — process*Node methods return
   ([]Resource, NodeState, error) instead of mutating shared *NodeState
   pointers. processNode becomes the single state registration point
   via SetNodeState.

6. Extract reconcileNodes pipeline — replace 90-line method with
   reconcileResult struct flowing through buildApplyInputs,
   applyResources, pruneIfSafe, and finalizeState phases.

All changes are internal (no public API changes). Build, vet, lint,
and all unit tests pass.
