+++ pkg/runtime/node.go
@@ -22,10 +22,8 @@ import (
 	"time"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apiserver/pkg/cel/openapi"
 	"k8s.io/kube-openapi/pkg/validation/spec"
 
-	celunstructured "github.com/kubernetes-sigs/kro/pkg/cel/unstructured"
 	"github.com/kubernetes-sigs/kro/pkg/graph"
 	"github.com/kubernetes-sigs/kro/pkg/graph/variable"
 	"github.com/kubernetes-sigs/kro/pkg/runtime/resolver"
@@ -60,6 +58,19 @@ var identityPaths = []string{
 	"metadata.name",
 	"metadata.namespace",
 }
+
+// resolveMode controls how template resolution behaves.
+type resolveMode int
+
+const (
+	// resolveFull evaluates all vars, fails on pending, caches result.
+	// Used by GetDesired for normal reconciliation.
+	resolveFull resolveMode = iota
+
+	// resolveIdentity evaluates identity paths only, no dep readiness check, no cache.
+	// Used by GetDesiredIdentity for deletion/observation.
+	resolveIdentity
+)
 
 // IsIgnored reports whether this node should be skipped entirely.
 // It is true when:
@@ -156,12 +167,27 @@ func (n *Node) IsIgnored() (bool, error) {
 //   - External: resolves template (for name/namespace CEL), caller reads instead of applies
 //
 // Note: The caller should call IsIgnored() before GetDesired() for resource nodes.
-func (n *Node) GetDesired() (result []*unstructured.Unstructured, err error) {
+func (n *Node) GetDesired() ([]*unstructured.Unstructured, error) {
 	// Return cached result if available.
 	if n.desired != nil {
 		return n.desired, nil
 	}
+	return n.resolve(resolveFull)
+}
 
+// GetDesiredIdentity resolves only identity-related fields (metadata.name & namespace)
+// and skips readiness gating. It is used for deletion/observation when we only need
+// stable identities and want to avoid being blocked by unrelated template fields.
+//
+// NOTE: This method does not cache its result in n.desired; callers in non-deletion
+// paths should continue using GetDesired().
+func (n *Node) GetDesiredIdentity() ([]*unstructured.Unstructured, error) {
+	return n.resolve(resolveIdentity)
+}
+
+// resolve is the unified resolution method. The mode controls which vars are
+// resolved, whether to tolerate pending dependencies, and whether to cache.
+func (n *Node) resolve(mode resolveMode) (result []*unstructured.Unstructured, err error) {
 	startTime := time.Now()
 	defer func() {
 		duration := time.Since(startTime)
@@ -172,9 +198,8 @@ func (n *Node) GetDesired() (result []*unstructured.Unstructured, err error) {
 		}
 	}()
 
-	// For resource types, block until all dependencies are ready.
-	// This enforces readyWhen semantics: dependents wait for parents.
-	if n.Spec.Meta.Type != graph.NodeTypeInstance {
+	// Full and partial modes check dep readiness (partial skips for instance nodes).
+	if mode == resolveFull && n.Spec.Meta.Type != graph.NodeTypeInstance {
 		for depID, dep := range n.deps {
 			if depID == graph.InstanceNodeID {
 				continue
@@ -189,106 +214,50 @@ func (n *Node) GetDesired() (result []*unstructured.Unstructured, err error) {
 		}
 	}
 
+	// Select vars based on mode.
+	vars := n.templateVars
+	if mode == resolveIdentity {
+		vars = n.templateVarsForPaths(identityPaths)
+	}
+
 	switch n.Spec.Meta.Type {
 	case graph.NodeTypeInstance:
+		if mode == resolveIdentity {
+			panic("GetDesiredIdentity called for instance node")
+		}
 		result, err = n.softResolve()
 	case graph.NodeTypeCollection:
-		result, err = n.hardResolveCollection(n.templateVars, true)
-	case graph.NodeTypeResource, graph.NodeTypeExternal:
-		// External refs resolve like resources (for name/namespace CEL),
-		// but the caller reads instead of applies.
-		result, err = n.hardResolveSingleResource(n.templateVars)
-	case graph.NodeTypeExternalCollection:
-		// Resolve the template to evaluate CEL expressions in
-		// metadata (name, namespace, selector). The caller extracts
-		// the resolved selector for LIST operations.
-		result, err = n.hardResolveSingleResource(n.templateVars)
-	default:
-		panic(fmt.Sprintf("unknown node type: %v", n.Spec.Meta.Type))
-	}
-
-	if err == nil {
-		if n.Spec.Meta.Type != graph.NodeTypeInstance {
-			if err = n.normalizeNamespaces(result); err != nil {
-				return nil, err
-			}
-		}
-		n.desired = result
-	}
-	return result, err
-}
-
-// GetDesiredIdentity resolves only identity-related fields (metadata.name & namespace)
-// and skips readiness gating. It is used for deletion/observation when we only need
-// stable identities and want to avoid being blocked by unrelated template fields.
-//
-// NOTE: This method does not cache its result in n.desired; callers in non-deletion
-// paths should continue using GetDesired().
-func (n *Node) GetDesiredIdentity() (result []*unstructured.Unstructured, err error) {
-	startTime := time.Now()
-	defer func() {
-		duration := time.Since(startTime)
-		nodeEvalDuration.Observe(duration.Seconds())
-		nodeEvalTotal.Inc()
-		if err != nil {
-			nodeEvalErrorsTotal.Inc()
-		}
-	}()
-
-	vars := n.templateVarsForPaths(identityPaths)
-	switch n.Spec.Meta.Type {
-	case graph.NodeTypeCollection:
-		result, err = n.hardResolveCollection(vars, false)
-		if err != nil {
-			return nil, err
-		}
-		if err := n.normalizeNamespaces(result); err != nil {
-			return nil, err
-		}
-		return result, nil
+		setIndexLabel := mode == resolveFull
+		result, err = n.hardResolveCollection(vars, setIndexLabel)
 	case graph.NodeTypeResource, graph.NodeTypeExternal:
 		result, err = n.hardResolveSingleResource(vars)
-		if err != nil {
-			return nil, err
-		}
-		if err := n.normalizeNamespaces(result); err != nil {
-			return nil, err
-		}
-		return result, nil
 	case graph.NodeTypeExternalCollection:
-		// External collections have no identity to resolve; they use selectors.
-		return nil, nil
-	case graph.NodeTypeInstance:
-		panic("GetDesiredIdentity called for instance node")
+		if mode == resolveIdentity {
+			// External collections have no identity to resolve; they use selectors.
+			return nil, nil
+		}
+		result, err = n.hardResolveSingleResource(vars)
 	default:
 		panic(fmt.Sprintf("unknown node type: %v", n.Spec.Meta.Type))
 	}
-}
 
-// normalizeNamespaces inherits the instance namespace onto namespaced children
-// that don't specify one. Cluster-scoped instances must resolve an explicit
-// namespace for namespaced children, otherwise reconciliation cannot safely
-// address them.
-func (n *Node) normalizeNamespaces(objs []*unstructured.Unstructured) error {
-	if !n.Spec.Meta.Namespaced {
-		return nil
+	if err != nil {
+		return nil, err
 	}
-	ns := n.deps[graph.InstanceNodeID].observed[0].GetNamespace()
-	for _, obj := range objs {
-		if obj.GetNamespace() != "" {
-			continue
-		}
-		if ns == "" {
-			return fmt.Errorf(
-				"node %q is namespaced and must resolve metadata.namespace when the instance is cluster-scoped",
-				n.Spec.Meta.ID,
-			)
-		}
-		if obj.GetNamespace() == "" {
-			obj.SetNamespace(ns)
+
+	// Normalize namespaces (not for instance nodes).
+	if n.Spec.Meta.Type != graph.NodeTypeInstance {
+		if err = n.normalizeNamespaces(result); err != nil {
+			return nil, err
 		}
 	}
-	return nil
+
+	// Cache result for full and partial modes (not identity).
+	if mode != resolveIdentity {
+		n.desired = result
+	}
+
+	return result, nil
 }
 
 // DeleteTargets returns the ordered list of objects this node should delete now.
@@ -318,247 +287,6 @@ func (n *Node) DeleteTargets() ([]*unstructured.Unstructured, error) {
 	}
 }
 
-func (n *Node) hardResolveSingleResource(vars []*variable.ResourceField) ([]*unstructured.Unstructured, error) {
-	baseExprs, _ := n.exprSetsForVars(vars)
-	values, _, err := n.evaluateExprsFiltered(baseExprs, false)
-	if err != nil {
-		return nil, fmt.Errorf("node %q: %w", n.Spec.Meta.ID, err)
-	}
-
-	desired := n.Spec.Template.DeepCopy()
-	res := resolver.NewResolver(desired.Object, values)
-	summary := res.Resolve(toFieldDescriptors(vars))
-	if len(summary.Errors) > 0 {
-		return nil, fmt.Errorf("node %q: resolve errors: %v", n.Spec.Meta.ID, summary.Errors)
-	}
-
-	return []*unstructured.Unstructured{desired}, nil
-}
-
-func (n *Node) hardResolveCollection(vars []*variable.ResourceField, setIndexLabel bool) ([]*unstructured.Unstructured, error) {
-	baseExprs, iterExprs := n.exprSetsForVars(vars)
-	baseValues, _, err := n.evaluateExprsFiltered(baseExprs, false)
-	if err != nil {
-		if !IsDataPending(err) {
-			err = fmt.Errorf("node %q base eval: %w", n.Spec.Meta.ID, err)
-		}
-		return nil, err
-	}
-
-	items, err := n.evaluateForEach()
-	if err != nil {
-		return nil, err
-	}
-
-	collectionSize.Observe(float64(len(items)))
-
-	if len(items) == 0 {
-		// Resolved empty collection: return non-nil empty slice to distinguish
-		// from unresolved (n.desired == nil).
-		return []*unstructured.Unstructured{}, nil
-	}
-
-	// Build a map from expression string to expressionEvaluationState for iteration expressions.
-	iterExprStates := make(map[string]*expressionEvaluationState, len(n.templateExprs))
-	for _, expr := range n.templateExprs {
-		if expr.Kind.IsIteration() {
-			iterExprStates[expr.Expression.Original] = expr
-		}
-	}
-
-	// Only build context for dependencies referenced by iteration expressions.
-	iterNeeded := make(map[string]struct{})
-	for exprStr := range iterExprs {
-		if state, ok := iterExprStates[exprStr]; ok {
-			for _, ref := range state.Expression.References {
-				iterNeeded[ref] = struct{}{}
-			}
-		}
-	}
-	baseCtx := n.buildContext(slices.Collect(maps.Keys(iterNeeded))...)
-
-	expanded := make([]*unstructured.Unstructured, 0, len(items))
-	for idx, iterCtx := range items {
-		values := make(map[string]any, len(baseValues)+len(iterExprs))
-		maps.Copy(values, baseValues)
-
-		// Merge iterator values into context.
-		ctx := make(map[string]any, len(baseCtx)+len(iterCtx))
-		maps.Copy(ctx, baseCtx)
-		maps.Copy(ctx, iterCtx)
-
-		// Evaluate iteration expressions (not cached - different context per iteration).
-		for exprStr := range iterExprs {
-			exprState := iterExprStates[exprStr]
-			val, err := exprState.Expression.Eval(ctx)
-			if err != nil {
-				if isCELDataPending(err) {
-					return nil, ErrDataPending
-				}
-				return nil, fmt.Errorf("collection iteration eval %q: %w", exprStr, err)
-			}
-			values[exprStr] = val
-		}
-
-		desired := n.Spec.Template.DeepCopy()
-		res := resolver.NewResolver(desired.Object, values)
-		summary := res.Resolve(toFieldDescriptors(vars))
-		if len(summary.Errors) > 0 {
-			return nil, fmt.Errorf("node %q collection resolve: resolve errors: %v", n.Spec.Meta.ID, summary.Errors)
-		}
-		if setIndexLabel {
-			setCollectionIndexLabel(desired, idx)
-		}
-		expanded = append(expanded, desired)
-	}
-
-	if err := validateUniqueIdentities(expanded); err != nil {
-		return nil, fmt.Errorf("node %q identity collision: %w", n.Spec.Meta.ID, err)
-	}
-
-	return expanded, nil
-}
-
-// softResolve evaluates expressions using best-effort partial resolution.
-// It ignores ErrDataPending (returns partial result) but propagates fatal errors.
-// Used for instance status where we populate as many fields as possible.
-//
-// Only fields where ALL expressions are resolved will be included in the result.
-// This prevents template strings like "${expr}" from leaking into the status.
-func (n *Node) softResolve() ([]*unstructured.Unstructured, error) {
-	values, _, err := n.evaluateExprsFiltered(nil, true) // soft: continue on pending
-	if err != nil {
-		return nil, err
-	}
-
-	// Filter to only fully-resolvable fields (expression value available)
-	var resolvable []variable.FieldDescriptor
-	for _, v := range n.templateVars {
-		if _, ok := values[v.Expression.Original]; ok {
-			resolvable = append(resolvable, v.FieldDescriptor)
-		}
-	}
-
-	// Resolve on template copy, then copy resolved values to empty desired
-	template := n.Spec.Template.DeepCopy()
-	templateRes := resolver.NewResolver(template.Object, values)
-	summary := templateRes.Resolve(resolvable)
-
-	// Resolution errors on filtered fields indicate bugs (template/path mismatch)
-	if len(summary.Errors) > 0 {
-		return nil, fmt.Errorf("failed to resolve status fields: %v", summary.Errors)
-	}
-
-	// Build desired with only successfully resolved fields
-	desired := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"status": map[string]interface{}{},
-		},
-	}
-	destRes := resolver.NewResolver(desired.Object, nil)
-	for _, result := range summary.Results {
-		if result.Resolved {
-			if err := destRes.UpsertValueAtPath(result.Path, result.Replaced); err != nil {
-				return nil, fmt.Errorf("failed to set status field %s: %w", result.Path, err)
-			}
-		}
-	}
-
-	return []*unstructured.Unstructured{desired}, nil
-}
-
-// evaluateExprsFiltered evaluates non-iteration expressions and returns the values map.
-// If exprs is nil, all expressions are evaluated. If exprs is empty, returns empty values.
-// If continueOnPending is true, it skips expressions that return ErrDataPending.
-// Returns (values, hasPending, error).
-func (n *Node) evaluateExprsFiltered(exprs map[string]struct{}, continueOnPending bool) (map[string]any, bool, error) {
-	if exprs != nil && len(exprs) == 0 {
-		return map[string]any{}, false, nil
-	}
-
-	// Compute the union of referenced dependencies across all expressions to
-	// evaluate, so buildContext only wraps needed deps with schema-aware values.
-	needed := n.neededDeps(exprs)
-	ctx := n.buildContext(needed...)
-
-	capacity := len(n.templateExprs)
-	if exprs != nil {
-		capacity = len(exprs)
-	}
-	values := make(map[string]any, capacity)
-	var hasPending bool
-	for _, expr := range n.templateExprs {
-		if expr.Kind.IsIteration() {
-			continue
-		}
-		if exprs != nil {
-			if _, ok := exprs[expr.Expression.Original]; !ok {
-				continue
-			}
-		}
-		if !expr.Resolved {
-			val, err := evalExprAny(expr, ctx)
-			if err != nil {
-				if isCELDataPending(err) {
-					hasPending = true
-					if continueOnPending {
-						continue
-					}
-					return nil, true, fmt.Errorf("failed to evaluate expression: %w (%w)", err, ErrDataPending)
-				}
-				return nil, false, err
-			}
-			expr.Resolved = true
-			expr.ResolvedValue = val
-		}
-		values[expr.Expression.Original] = expr.ResolvedValue
-	}
-	return values, hasPending, nil
-}
-
-func (n *Node) templateVarsForPaths(paths []string) []*variable.ResourceField {
-	if len(paths) == 0 {
-		return n.templateVars
-	}
-
-	pathSet := make(map[string]struct{}, len(paths))
-	for _, p := range paths {
-		pathSet[p] = struct{}{}
-	}
-
-	result := make([]*variable.ResourceField, 0, len(n.templateVars))
-	for _, v := range n.templateVars {
-		if _, ok := pathSet[v.Path]; ok {
-			result = append(result, v)
-		}
-	}
-	return result
-}
-
-func (n *Node) exprSetsForVars(
-	vars []*variable.ResourceField,
-) (map[string]struct{}, map[string]struct{}) {
-	baseExprs := make(map[string]struct{})
-	iterExprs := make(map[string]struct{})
-	if len(vars) == 0 {
-		return baseExprs, iterExprs
-	}
-
-	exprKinds := make(map[string]variable.ResourceVariableKind, len(n.templateExprs))
-	for _, expr := range n.templateExprs {
-		exprKinds[expr.Expression.Original] = expr.Kind
-	}
-
-	for _, v := range vars {
-		if kind, ok := exprKinds[v.Expression.Original]; ok && kind.IsIteration() {
-			iterExprs[v.Expression.Original] = struct{}{}
-		} else {
-			baseExprs[v.Expression.Original] = struct{}{}
-		}
-	}
-	return baseExprs, iterExprs
-}
-
 // upsertToTemplate applies values by upserting at paths, creating parent fields if needed.
 // Use for instance status where paths like status.foo may not exist yet.
 func (n *Node) upsertToTemplate(base *unstructured.Unstructured, values map[string]any) *unstructured.Unstructured {
@@ -584,248 +312,4 @@ func (n *Node) SetObserved(observed []*unstructured.Unstructured) {
 	default:
 		n.observed = observed
 	}
-}
-
-// CheckReadiness evaluates readyWhen expressions using observed state.
-// Ignored nodes are treated as ready for dependency gating purposes.
-func (n *Node) CheckReadiness() error {
-	nodeReadyCheckTotal.Inc()
-
-	// Ignored nodes are satisfied for dependency gating - dependents shouldn't block.
-	ignored, err := n.IsIgnored()
-	if err != nil {
-		return fmt.Errorf("is ignore check failed: %w", err)
-	}
-	if ignored {
-		return nil
-	}
-
-	err = n.checkObservedReadiness()
-	if err != nil && errors.Is(err, ErrWaitingForReadiness) {
-		nodeNotReadyTotal.Inc()
-	}
-
-	return err
-}
-
-// checkObservedReadiness evaluates readiness from the node's own observed and
-// desired state. Callers that need ignore semantics must handle them first.
-func (n *Node) checkObservedReadiness() error {
-	if n.Spec.Meta.Type == graph.NodeTypeCollection || n.Spec.Meta.Type == graph.NodeTypeExternalCollection {
-		return n.checkCollectionReadiness()
-	}
-	return n.checkSingleResourceReadiness()
-}
-
-func (n *Node) checkSingleResourceReadiness() error {
-	if len(n.observed) == 0 {
-		return fmt.Errorf("node %q: no observed state: %w", n.Spec.Meta.ID, ErrWaitingForReadiness)
-	}
-	if len(n.readyWhenExprs) == 0 {
-		return nil
-	}
-
-	nodeID := n.Spec.Meta.ID
-	ctx := map[string]any{nodeID: n.observed[0].Object}
-
-	for _, expr := range n.readyWhenExprs {
-		result, err := evalBoolExpr(expr, ctx)
-		if err != nil {
-			if isCELDataPending(err) {
-				return fmt.Errorf("node %q: failed to evaluate readyWhen expression: %q (%w)", n.Spec.Meta.ID, expr.Expression.UserExpression(), ErrWaitingForReadiness)
-			}
-			return fmt.Errorf("node %q: failed to evaluate readyWhen expression: %q (%w)", n.Spec.Meta.ID, expr.Expression.UserExpression(), err)
-		}
-		if !result {
-			return fmt.Errorf("readyWhen condition evaluated to false: %q (%w)", expr.Expression.UserExpression(), ErrWaitingForReadiness)
-		}
-	}
-	return nil
-}
-
-func (n *Node) checkCollectionReadiness() error {
-	if n.Spec.Meta.Type == graph.NodeTypeExternalCollection {
-		// External collections: desired carries the selector template, not actual
-		// desired resources. Skip count-based readiness checks.
-		if len(n.readyWhenExprs) == 0 || len(n.observed) == 0 {
-			return nil
-		}
-	} else {
-		// Use nil check (not len==0) to distinguish "not computed" from "empty collection".
-		if n.desired == nil {
-			return fmt.Errorf("node %q: collection not computed (%w)", n.Spec.Meta.ID, ErrWaitingForReadiness)
-		}
-		if len(n.desired) == 0 {
-			return nil
-		}
-		if len(n.observed) < len(n.desired) {
-			return fmt.Errorf("node %q: collection not ready: observed %d but desired %d (%w)", n.Spec.Meta.ID, len(n.observed), len(n.desired), ErrWaitingForReadiness)
-		}
-		if len(n.readyWhenExprs) == 0 {
-			return nil
-		}
-	}
-
-	// Collection readyWhen uses "each" (single item) only.
-	// Each item has different context, so we evaluate directly (not cached).
-	for i, obj := range n.observed {
-		ctx := map[string]any{graph.EachVarName: obj.Object}
-		for _, expr := range n.readyWhenExprs {
-			// readyWhen for collections must NOT be cached - each item has different "each" context.
-			// Use Expression.Eval directly instead of evalBoolExpr.
-			val, err := expr.Expression.Eval(ctx)
-			if err != nil {
-				if isCELDataPending(err) {
-					return fmt.Errorf("node %q: failed to evaluate readyWhen %q (item %d) (%w)", n.Spec.Meta.ID, expr.Expression.UserExpression(), i, ErrWaitingForReadiness)
-				}
-				return fmt.Errorf("node %q: failed to evaluate readyWhen %q (item %d): %w", n.Spec.Meta.ID, expr.Expression.UserExpression(), i, err)
-			}
-			result, ok := val.(bool)
-			if !ok {
-				return fmt.Errorf("readyWhen %q did not return bool", expr.Expression.UserExpression())
-			}
-			if !result {
-				return fmt.Errorf("readyWhen condition evaluated to false: %q (%w)", expr.Expression.UserExpression(), ErrWaitingForReadiness)
-			}
-		}
-	}
-
-	return nil
-}
-
-// evaluateForEach evaluates forEach dimensions and returns iterator contexts.
-func (n *Node) evaluateForEach() ([]map[string]any, error) {
-	if len(n.Spec.ForEach) == 0 {
-		return nil, nil
-	}
-
-	// Only build context for dependencies referenced by forEach expressions.
-	needed := make(map[string]struct{})
-	for _, expr := range n.forEachExprs {
-		for _, ref := range expr.Expression.References {
-			needed[ref] = struct{}{}
-		}
-	}
-	ctx := n.buildContext(slices.Collect(maps.Keys(needed))...)
-
-	dimensions := make([]evaluatedDimension, len(n.Spec.ForEach))
-	for i, dim := range n.Spec.ForEach {
-		values, err := evalListExpr(n.forEachExprs[i], ctx)
-		if err != nil {
-			if isCELDataPending(err) {
-				return nil, ErrDataPending
-			}
-			return nil, fmt.Errorf("forEach %q: %w", dim.Name, err)
-		}
-		if len(values) == 0 {
-			return nil, nil
-		}
-		dimensions[i] = evaluatedDimension{name: dim.Name, values: values}
-	}
-
-	product, err := cartesianProduct(dimensions, n.rgdConfig.MaxCollectionSize)
-	if err != nil {
-		return nil, err
-	}
-
-	return product, nil
-}
-
-// buildContext builds the CEL activation context from node dependencies.
-// If only is provided, only those dependency IDs are included in the context.
-// If only is empty/nil, all dependencies are included.
-//
-// When a dependency has a resourceSchema, its observed objects are wrapped using
-// Kubernetes' UnstructuredToVal for schema-aware type conversion. This ensures
-// CEL runtime values match their compile-time types (e.g., Secret data as bytes).
-func (n *Node) buildContext(only ...string) map[string]any {
-	ctx := make(map[string]any)
-	for depID, dep := range n.deps {
-		// Use nil check (not len==0) to include empty collections in context.
-		if dep.observed == nil {
-			continue
-		}
-		if len(only) > 0 && !slices.Contains(only, depID) {
-			continue
-		}
-		if dep.Spec.Meta.Type == graph.NodeTypeCollection || dep.Spec.Meta.Type == graph.NodeTypeExternalCollection {
-			items := make([]any, len(dep.observed))
-			for i, obj := range dep.observed {
-				items[i] = wrapWithSchema(obj.Object, dep.resourceSchema)
-			}
-			ctx[depID] = items
-		} else {
-			obj := dep.observed[0].Object
-			// For schema (instance), strip status - users should only access spec/metadata.
-			// The instance's resourceSchema already excludes status (set by builder via
-			// getSchemaWithoutStatus), so the schema and data stay aligned.
-			if depID == graph.InstanceNodeID {
-				obj = withStatusOmitted(obj)
-			}
-			ctx[depID] = wrapWithSchema(obj, dep.resourceSchema)
-		}
-	}
-	return ctx
-}
-
-// wrapWithSchema wraps an unstructured object with schema-aware CEL value
-// conversion. If the schema is nil, the raw object is returned. Otherwise,
-// returns a schemaMap that delegates to UnstructuredToVal for typed properties
-// and falls back to NativeToValue for preserve-unknown fields.
-func wrapWithSchema(obj map[string]interface{}, schema *spec.Schema) any {
-	if schema == nil {
-		return obj
-	}
-	return celunstructured.UnstructuredToVal(obj, &openapi.Schema{Schema: schema})
-}
-
-// withStatusOmitted returns a shallow copy of obj with the "status" key removed.
-// This prevents CEL expressions from accessing instance status fields.
-func withStatusOmitted(obj map[string]any) map[string]any {
-	result := make(map[string]any, len(obj))
-	for k, v := range obj {
-		if k != "status" {
-			result[k] = v
-		}
-	}
-	return result
-}
-
-// neededDeps computes the union of referenced dependency IDs across the
-// expressions in the given set. If exprs is nil, all non-iteration template
-// expressions are included. This allows buildContext to only wrap needed deps.
-func (n *Node) neededDeps(exprs map[string]struct{}) []string {
-	needed := make(map[string]struct{})
-	for _, expr := range n.templateExprs {
-		if expr.Kind.IsIteration() {
-			continue
-		}
-		if exprs != nil {
-			if _, ok := exprs[expr.Expression.Original]; !ok {
-				continue
-			}
-		}
-		for _, ref := range expr.Expression.References {
-			needed[ref] = struct{}{}
-		}
-	}
-	return slices.Collect(maps.Keys(needed))
-}
-
-// contextDependencyIDs returns CEL variable names grouped by type.
-// - singles: dependencies that are single resources (declared as dyn)
-// - collections: dependencies that are collections (declared as list(dyn))
-// - iterators: forEach loop variable names from iterCtx (declared as list(dyn))
-func (n *Node) contextDependencyIDs(iterCtx map[string]any) (singles, collections, iterators []string) {
-	for depID, dep := range n.deps {
-		if dep.Spec.Meta.Type == graph.NodeTypeCollection || dep.Spec.Meta.Type == graph.NodeTypeExternalCollection {
-			collections = append(collections, depID)
-		} else {
-			singles = append(singles, depID)
-		}
-	}
-	for name := range iterCtx {
-		iterators = append(iterators, name)
-	}
-	return
 }
